untuk selanjutnya adalah fitur delete dari operator

+++ web_app/form.py
@@ -1,6 +1,6 @@
 from wtforms.ext.sqlalchemy.fields import QuerySelectField
 from wtforms.widgets import TextArea
-from wtforms import StringField, PasswordField, BooleanField, TextAreaField, IntegerField, SelectField, DateTimeField
+from wtforms import StringField, PasswordField, BooleanField, TextAreaField, IntegerField, SelectField, DateTimeField, DecimalField
 from wtforms.validators import InputRequired, Email, Length, DataRequired, Required
 from flask_wtf import FlaskForm
 from wtforms.fields.html5 import DateField
@@ -36,9 +36,9 @@ class SocietyInputDataView(FlaskForm):
 
 
 class OperatorAddDataView(FlaskForm):
-    NIK_catin_laki_laki = IntegerField('', validators=[DataRequired()], render_kw={'placeholder': 'NIK Catin Laki-laki'})
+    NIK_catin_laki_laki = DecimalField('', validators=[DataRequired()], render_kw={'placeholder': 'NIK Catin Laki-laki'})
     nama_catin_laki_laki = StringField('', validators=[DataRequired()], render_kw={"placeholder": "Nama Catin Laki-laki"})
-    NIK_catin_perempuan = IntegerField('', validators=[DataRequired()], render_kw={'placeholder': 'NIK Catin Perempuan'})
+    NIK_catin_perempuan = DecimalField('', validators=[DataRequired()], render_kw={'placeholder': 'NIK Catin Perempuan'})
     nama_catin_perempuan = StringField('', validators=[DataRequired()], render_kw={"placeholder": "Nama Catin Perempuan"})
     jadwal_nikah = DateField('Jadwal Nikah', format='%Y-%m-%d', validators=[DataRequired()])
     tempat_pelaksaan_nikah = StringField('', validators=[DataRequired()], render_kw={"placeholder": "Tempat Pelaksanaan Nikah"})
